Refactor heap bindings to use UseOurHeap and FixHeap

+++ shim/bindings_ntdll.h
@@ -5,7 +5,7 @@
 typedef PVOID NTAPI tRtlAllocateHeap(IN HANDLE HeapHandle, IN ULONG Flags, IN SIZE_T Size);
 typedef PVOID NTAPI tRtlReAllocateHeap(IN HANDLE HeapHandle, IN ULONG Flags, IN LPVOID Address, IN SIZE_T Size);
 typedef BOOLEAN NTAPI tRtlFreeHeap(IN HANDLE HeapHandle, IN ULONG Flags, IN PVOID Address);
-typedef ULONG NTAPI tRtlSizeHeap(IN HANDLE HeapHandle, IN ULONG Flags, IN PVOID Address);
+typedef SIZE_T NTAPI tRtlSizeHeap(IN HANDLE HeapHandle, IN ULONG Flags, IN PVOID Address);
 
 extern tRtlAllocateHeap* o_RtlAllocateHeap;
 extern tRtlReAllocateHeap* o_RtlReAllocateHeap;
@@ -15,4 +15,4 @@ extern tRtlSizeHeap* o_RtlSizeHeap;
 PVOID NTAPI h_RtlAllocateHeap(IN HANDLE HeapHandle, IN ULONG Flags, IN SIZE_T Size);
 PVOID NTAPI h_RtlReAllocateHeap(IN HANDLE HeapHandle, IN ULONG Flags, IN LPVOID Address, IN SIZE_T Size);
 BOOLEAN NTAPI h_RtlFreeHeap(IN HANDLE HeapHandle, IN ULONG Flags, IN PVOID Address);
-ULONG NTAPI h_RtlSizeHeap(IN HANDLE HeapHandle, IN ULONG Flags, IN PVOID Address);
+SIZE_T NTAPI h_RtlSizeHeap(IN HANDLE HeapHandle, IN ULONG Flags, IN PVOID Address);
